# Optimize *.test_sse4_1 test runtime.  NFC