LLVM struggles with compilation time on large functions, and counterproductively inlines functions to make things worse for itself. Outline each SSE function to be tested in separate noninlineable function.

Before: Total core time: 487.852s. Wallclock time: 323.379s. Parallelization: 1.51x.
After: Total core time: 243.685s. Wallclock time: 120.836s. Parallelization: 2.02x.